requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
tqdm==4.31.1
urllib3==1.24.1
wget==3.2